backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-yasg==1.21.7
Pillow==10.4.0
PyJWT==2.8.0
python-dotenv==1.0.1
gunicorn==22.0.0
